[BUILD] require Stmt.java, Expr.java for all files
because java allows cyclic imports, it's impossible to write a makefile
for an arbitrary java project. the previous method of 'compile all the
class files' works because all the source files existed. but `javac`
looks at the newest file of .java and .class:
https://docs.oracle.com/javase/8/docs/technotes/tools/windows/javac.html#BHCCHDGH
and some classes need to be compiled before others. But some classes depend
on each other, for example in this project:

Lexer.java <- Lox.java <- Lexer.java

The only real way to deal with this is to compile all the top-level
roots of the tree with `javac` and let it figure out the rest.

In this project, the current top-level roots are:
- Lox.java, which depends on every file but
- Compiler.java

The problem I encountered (and fixed in this commit) is that I
autogenerate java sources, so `javac` can't find some of the files. This
requires the makefile to have knowledge of the build. Although this is
impossible in the general case (as stated above), the files Expr.java
and Stmt.java have no dependencies (besides the standard library), so we
can special case them.

This is the long way round of saying every .java file now requires
Expr.java and Stmt.java to exist before `javac` runs.